refactor: GeneralTest -> CastleMovementTest로 일반화
